AUDIT-53:Moved ReadAuditAOP helper method to different component class,and securing it.
